[3.12] gh-122292: Split up ``Lib/test/test_ast.py`` into a couple of files (GH-122293) (cherry picked from commit 9187484dd97f6beb94fc17676014706922e380e1)
